apps/arcs/sigma-agent/requirements.txt
nats-py==2.6.0
uvloop==0.19.0
orjson==3.10.7
